pyproject.toml
[tool.poetry]
name = "qtgql"
version = "0.119.13"
packages = [{ include = "qtgqlcodegen" }]
description = "Qt framework for building graphql driven QML applications"
authors = ["Nir <88795475+nrbnlulu@users.noreply.github.com>"]
maintainers = ["Nir.J Benlulu <nrbnlulu@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.urls]
"Homepage" = "https://github.com/nrbnlulu/qtgql"
"Documentation" = "https://nrbnlulu.github.io/qtgql/"


[tool.poetry.dependencies]
python = ">=3.9,<3.12"
attrs = "^22.2.0"
# codegen
graphql-core = "^3.2.3"
jinja2 = {version = "^3.1.2", optional=true}
typingref = "^0.100.0"
typer = "^0.9.0"
rich = "^13.3.5"
frozendict = "^2.3.8"



[tool.poetry.group.dev.dependencies]
aiohttp = {extras = ["speedups"], version = "^3.8.3"}
faker = ">=15.3.4,<19.0.0"
mypy = ">=1.0.1"
strawberry-graphql = ">=0.158.2"
aqtinstall = "v3.1.6"
conan = ">=2.0.4"
cmake = ">=3.25.0"
pygithub = ">=1.58.2"
githubrelease = ">=1.5.9"
httpx = ">=0.18.2"
autopub = ">=0.2.2"
mimesis = ">=10.1.0"
pytest-cov = ">=4.1.0"


[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.0.5"
mike = "^1.1.2"
markdown-include = "^0.8.0"
mkdocstrings = {extras = ["python"], version = ">=0.19.1,<0.22.0"}
pygments = "^2.14.0"






[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.mypy]
ignore_missing_imports = "True"
exclude = "(?x)^(.*\\.jinja.py|)$"
warn_no_return = "False"

[tool.black]
line-length = 100
target-version = ['py310']
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | __pycached__
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
select = ["ALL"]
target-version = "py38"
ignore = [
    "TID252",
    # we use asserts in tests and to hint mypy
    "E501", # line too long, handled by black.
    "S101",
    "S102",
    "S104",
    "S324",
    "EXE002",
    # maybe we can enable this in future
    # we'd want to have consistent docstrings in future
    "D",
    "ANN101", # missing annotation for self?
    # definitely enable these, maybe not in tests
    "ANN001",
    "ANN002",
    "ANN003",
    "ANN102",
    "ANN201",
    "ANN202",
    "ANN204",
    "ANN205",
    "ANN206",
    "ANN401",
    "PGH003",
    "PGH004",
    "RET504",
    "RET505",
    "RET506",
    "RET507",
    "BLE001",
    "RET503",
    "B008",
    "N811",
    "N804",
    "N818",
    # Variable `T` in function should be lowercase
    # this seems a potential bug or opportunity for improvement in ruff
    "N806",


    "N815",

    # shadowing builtins
    "A001",
    "A002",
    "A003",

    "ARG001",
    "ARG002",
    "ARG003",
    "ARG004",
    "ARG005",
    "FBT001",
    "FBT002",
    "FBT003",

    "PT001",
    "PT023",

    # enable these, we have some in tests
    "B006",
    "PT004",
    "PT007",
    "PT011",
    "PT012",
    "PT015",
    "PT017",
    "C414",
    "N802",

    "SIM117",
    "SIM102",

    "F841",
    "B027",
    "B905",
    "ISC001",

    # same?
    "S105",
    "S106",

    "DTZ003",
    "DTZ005",

    "RSE102",
    "SLF001",

    # in tests
    "DTZ001",

    "EM101",
    "EM102",
    "EM103",

    "B904",
    "B019",

    "N801",
    "N807",

    # pandas
    "PD",

    # code complexity
    "C",
    "C901",

    # trailing commas
    "COM812",

    "PLR",
    "INP",
    "TRY",
    "SIM300",
    "SIM114",
    "DJ008",

    # https://beta.ruff.rs/docs/rules/#flake8-fixme-fix
    "FIX002",
    # https://beta.ruff.rs/docs/rules/#flake8-bandit-s
    "S603",
    "S607",

]
fix = true
src = ["qtgqlcodegen", "tests"]
[tool.ruff.flake8-annotations]
suppress-none-returning = true

[tool.ruff.isort]
combine-as-imports = true

[tool.pytest.ini_options]
addopts = "--cov-config=pyproject.toml --ignore=qtgqlcodegen/codegen/py/templates/"


[tool.coverage.run]
source = ["qtgqlcodegen"]
omit = ['test_*', 'tests/*', '**/*.jinja.py']
relative_files = true

[tool.coverage.report]
show_missing = true
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "def __str__",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "if typing.TYPE_CHECKING:"
]

[tool.qtgql]
config = "tests.test_sample_ui.main:qtgqlconfig"


[tool.autopub]
git-username = "QtGqlBot"
git-email = "bot@no.reply"
project-name = "qtgql"
append-github-contributor = true

[tool.poetry.scripts]
qtgql = "qtgql.cli:entrypoint"
